+++ changelog.md
@@ -1,4 +1,3 @@
 ### General
 
-- ***Added*** carpet rule `tickFluids`
-- Added command `/ad xaero <id>` to share address as Xaero waypoint
+- Fixed improper rule categories